First draft solution
This is my first solution to the problem where I didn't realise that the array numbers was just meant to be a list of numbers from 1 to n; rather I thought it was a set of random numbers from 1 to n. I also realised this isn't the most efficient of programs. I am now working to find a solution with linear time complexity and constant space.